Added account name to profile page
I added an account name and account image to the profile page.

The account name is customizable through a variable which will be received by the server when we work on the back end.

I still have to find out how to make the picture customizable though.

How to use:
<Profile accountName='Put account name' />